Try moving admin-sidebar out of initializer
