+++ app/assets/javascripts/discourse/app/lib/sidebar/admin-sidebar.js
@@ -1,10 +1,12 @@
+import { getOwner } from "@ember/application";
 import PreloadStore from "discourse/lib/preload-store";
 import { ADMIN_NAV_MAP } from "discourse/lib/sidebar/admin-nav-map";
-import {
-  addSidebarPanel,
-  addSidebarSection,
-} from "discourse/lib/sidebar/custom-sections";
-import { ADMIN_PANEL } from "discourse/services/sidebar-state";
+import BaseCustomSidebarPanel from "discourse/lib/sidebar/base-custom-sidebar-panel";
+import BaseCustomSidebarSection from "discourse/lib/sidebar/base-custom-sidebar-section";
+import BaseCustomSidebarSectionLink from "discourse/lib/sidebar/base-custom-sidebar-section-link";
+// import { ADMIN_PANEL } from "discourse/services/sidebar-state";
+const ADMIN_PANEL = "admin"; // TODO... importing from services/sidebar-state causes a circular module dependency
+import { getOwnerWithFallback } from "discourse-common/lib/get-owner";
 import I18n from "discourse-i18n";
 
 let additionalAdminSidebarSectionLinks = {};
@@ -14,8 +16,8 @@ export function clearAdditionalAdminSidebarSectionLinks() {
   additionalAdminSidebarSectionLinks = {};
 }
 
-function defineAdminSectionLink(BaseCustomSidebarSectionLink) {
-  const SidebarAdminSectionLink = class extends BaseCustomSidebarSectionLink {
+function defineAdminSectionLink() {
+  class SidebarAdminSectionLink extends BaseCustomSidebarSectionLink {
     constructor({ adminSidebarNavLink }) {
       super(...arguments);
       this.adminSidebarNavLink = adminSidebarNavLink;
@@ -58,7 +60,7 @@ function defineAdminSectionLink(BaseCustomSidebarSectionLink) {
     get title() {
       return this.adminSidebarNavLink.text;
     }
-  };
+  }
 
   return SidebarAdminSectionLink;
 }
@@ -164,25 +166,6 @@ export function useAdminNavConfig(navMap) {
 }
 
 let adminSectionLinkClass = null;
-export function buildAdminSidebar(navConfig) {
-  navConfig.forEach((adminNavSectionData) => {
-    addSidebarSection(
-      (BaseCustomSidebarSection, BaseCustomSidebarSectionLink) => {
-        // We only want to define the link class once even though we have many different sections.
-        adminSectionLinkClass =
-          adminSectionLinkClass ||
-          defineAdminSectionLink(BaseCustomSidebarSectionLink);
-
-        return defineAdminSection(
-          adminNavSectionData,
-          BaseCustomSidebarSection,
-          adminSectionLinkClass
-        );
-      },
-      ADMIN_PANEL
-    );
-  });
-}
 
 // This is used for a plugin API.
 export function addAdminSidebarSectionLink(sectionName, link) {
@@ -237,33 +220,13 @@ function pluginAdminRouteLinks() {
   );
 }
 
-export default {
-  name: "admin-sidebar-initializer",
+export default class AdminSidebarPanel extends BaseCustomSidebarPanel {
+  key = ADMIN_PANEL;
+  hidden = true;
 
-  initialize(owner) {
-    this.currentUser = owner.lookup("service:current-user");
-    this.siteSettings = owner.lookup("service:site-settings");
-
-    if (
-      !this.currentUser?.staff ||
-      !this.siteSettings.userInAnyGroups(
-        "admin_sidebar_enabled_groups",
-        this.currentUser
-      )
-    ) {
-      return;
-    }
-
-    this.adminSidebarExperimentStateManager = owner.lookup(
+  get sections() {
+    this.adminSidebarExperimentStateManager = getOwnerWithFallback(this).lookup(
       "service:admin-sidebar-experiment-state-manager"
-    );
-
-    addSidebarPanel(
-      (BaseCustomSidebarPanel) =>
-        class AdminSidebarPanel extends BaseCustomSidebarPanel {
-          key = ADMIN_PANEL;
-          hidden = true;
-        }
     );
 
     const savedConfig = this.adminSidebarExperimentStateManager.navConfig;
@@ -271,7 +234,10 @@ export default {
 
     navMap.findBy("name", "plugins").links.push(...pluginAdminRouteLinks());
 
-    if (this.siteSettings.experimental_form_templates) {
+    if (
+      getOwnerWithFallback().lookup("service:site-settings")
+        .experimental_form_templates
+    ) {
       navMap.findBy("name", "customize").links.push({
         name: "admin_customize_form_templates",
         route: "adminCustomizeFormTemplates",
@@ -280,6 +246,17 @@ export default {
       });
     }
 
-    buildAdminSidebar(useAdminNavConfig(navMap), adminSectionLinkClass);
-  },
-};
+    const navConfig = useAdminNavConfig(navMap);
+
+    return navConfig.map((adminNavSectionData) => {
+      // We only want to define the link class once even though we have many different sections.
+      adminSectionLinkClass = adminSectionLinkClass || defineAdminSectionLink();
+
+      return defineAdminSection(
+        adminNavSectionData,
+        BaseCustomSidebarSection,
+        adminSectionLinkClass
+      );
+    });
+  }
+}
